% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getAsbestosFile}
\alias{getAsbestosFile}
\title{Function to import data (for Code4Health)}
\usage{
getAsbestosFile(dsn = NA, layer = NA, sep = c(";", ","), header = TRUE,
  force.spatial = TRUE, ...)
}
\arguments{
\item{dsn}{input file with data. See \code{\link{readOGR}}}

\item{layer}{see \code{\link{readOGR}}}

\item{sep}{potential separator characters for CSV files. See \code{\link{read.table}}}

\item{header}{See \code{\link{read.table}}}
}
\description{
Function to import data (for Code4Health)
}
\note{
It reads single source files in SHP, CSV or geoJSON format.
}
\author{
Emanuele Cordano, Spaghetti Open Data
}

